move trait to namespace
